moved status filter logic functions to separate file
