+++ frontend/src/components/App/index.js
@@ -3,6 +3,7 @@ import './styles.css';
 import FilterContainer from '../FilterContainer'
 import LocationView from '../LocationView'
 import GoogleMaps from '../GoogleMaps'
+import * as statusLogic from '../StatusFilterLogic'
 
 const aaltoLocations = require('../../api/aalto-with-trashbins.json')
 
@@ -44,9 +45,10 @@ class App extends Component {
       }
     })
 
-    const overflowLocations = this.getOverflowLocations(locations);
-    const overdueLocations = this.getOverdueLocations(locations);
-    const noIssueLocations = this.getNoIssueLocations(locations);
+    // filter additive functionality
+    const overflowLocations = statusLogic.getOverflowLocations(locations);
+    const overdueLocations = statusLogic.getOverdueLocations(locations);
+    const noIssueLocations = statusLogic.getNoIssueLocations(locations);
     const oveflowMap = overflowLocations.map(a => a.id);
     const overdueMap = overdueLocations.map(a => a.id);
     const issueMap = noIssueLocations.map(a => a.id);
@@ -55,13 +57,13 @@ class App extends Component {
       return locations;
     }
     if (statusFilters.get('Trash overflows') && statusFilters.get('Late pickups')) {
-      return locations.filter(a => this.arrUnion(oveflowMap, overdueMap).includes(a.id));
+      return locations.filter(a => statusLogic.arrUnion(oveflowMap, overdueMap).includes(a.id));
     }
     if (statusFilters.get('Trash overflows') && statusFilters.get('No issues')) {
-      return locations.filter(a => this.arrUnion(oveflowMap, issueMap).includes(a.id));
+      return locations.filter(a => statusLogic.arrUnion(oveflowMap, issueMap).includes(a.id));
     }
     if (statusFilters.get('Late pickups') && statusFilters.get('No issues')) {
-      return locations.filter(a => this.arrUnion(issueMap, overdueMap).includes(a.id));
+      return locations.filter(a => statusLogic.arrUnion(issueMap, overdueMap).includes(a.id));
     }
     if (statusFilters.get('Trash overflows')) {
       return overflowLocations;
@@ -73,113 +75,6 @@ class App extends Component {
       return noIssueLocations;
     }
     return locations;
-  }
-
-  getOverdueLocations(locations) {
-    return locations.filter(a => a.trashbins.filter(c => c.pickupOverdue === true).length !== 0 );
-  }
-
-  // Operate arrays using Set theory, https://en.wikipedia.org/wiki/Set_theory
-  getNoIssueLocations(locations) {
-    const origLocations = locations;
-    const overdueLocations = this.getOverdueLocations(locations);
-    const overflowLocations = this.getOverflowLocations(locations);
-    const mapAll = origLocations.map(a => a.id);
-    const mapOverdue = overdueLocations.map(a => a.id);
-    const mapOverflow = overflowLocations.map(a => a.id);
-
-    // i.e [1,2,3]-[1,2]=[3]
-    function aMinusB(a, b) {
-      return a.filter(
-        c => b.indexOf(c) < 0,
-      );
-    }
-
-    /*
-    We assume that no issue locations should be locations that have no overdue nor overflow status,
-    thus we remove the union of overflow and overdue ids from the list of all ids
-    to get the ids with no issues.
-      All-([Overdue]u[Overflow]) <-> [1,2,3,4,5]-([1,2]u[2,3]) = [4,5]
-    */
-    const noIssueIds = aMinusB(mapAll, (this.arrUnion(mapOverdue, mapOverflow)));
-
-    return locations.filter(a => noIssueIds.includes(a.id));
-  }
-
-  getOverflowLocations(locations) {
-    const overflowLocations = [];
-
-    locations.forEach((loc, i) => {
-      // const trashbins = loc.trashbins
-      const { trashbins } = loc;
-
-      trashbins.sort((a, b) => {
-        if (a.wasteType < b.wasteType) { return -1; }
-        if (a.wasteType > b.wasteType) { return 1; }
-        return 0;
-      })
-
-      let currentWasteType = trashbins[ 0 ].wasteType;
-      let binCounter = 0;
-      let fullCounter = 0;
-
-      trashbins.forEach((bin, j) => {
-        if (currentWasteType !== bin.wasteType) {
-          if (binCounter === fullCounter) {
-            overflowLocations.push(locations[ i ])
-          }
-          binCounter = 0;
-          fullCounter = 0;
-        }
-
-        currentWasteType = bin.wasteType
-        if (bin.fillStatus === 100) fullCounter += 1;
-        binCounter += 1;
-
-        if (j === trashbins.length - 1 && binCounter === fullCounter) {
-          overflowLocations.push(locations[ i ])
-        }
-      });
-    })
-
-    const uniqueOverflowLocations = [ ...new Set(overflowLocations) ];
-    return uniqueOverflowLocations;
-  }
-
-  getOverflowTypes(location) {
-    const overflowTypes = [];
-    // const trasbins = location.trashbins
-    const { trashbins } = location;
-
-    trashbins.sort((a, b) => {
-      if (a.wasteType < b.wasteType) { return -1; }
-      if (a.wasteType > b.wasteType) { return 1; }
-      return 0;
-    })
-
-    let currentWasteType = trashbins[ 0 ].wasteType;
-    let binCounter = 0;
-    let fullCounter = 0;
-
-    trashbins.forEach((bin, j) => {
-      if (currentWasteType !== bin.wasteType) {
-        if (binCounter === fullCounter) {
-          overflowTypes.push(currentWasteType)
-        }
-        binCounter = 0;
-        fullCounter = 0;
-      }
-
-      currentWasteType = bin.wasteType
-      if (bin.fillStatus === 100) fullCounter += 1;
-      binCounter += 1;
-
-      if (j === trashbins.length - 1 && binCounter === fullCounter) {
-        overflowTypes.push(currentWasteType)
-      }
-    });
-
-    return overflowTypes;
   }
 
   getSidebarView() {
@@ -201,22 +96,6 @@ class App extends Component {
             statusFilters={ statusFilters }
           />
       )
-  }
-
-  /*
-  Helpervfunction to create unions of arrays using Set theory
-  i.e [1,2,3]u[2,3,5] = [1,2,3,5]
-  */
-  arrUnion(a, b) {
-    const obj = {};
-    for (let i = a.length - 1; i >= 0; i -= 1) { obj[ a[ i ] ] = a[ i ]; }
-    for (let j = b.length - 1; j >= 0; j -= 1) { obj[ b[ j ] ] = b[ j ]; }
-    const res = []
-    // for (const k in obj) {
-    Object.keys(obj).forEach((key) => {
-      if (Object.prototype.hasOwnProperty.call(obj, key)) { res.push(obj[ key ]); }
-    })
-    return res;
   }
 
   /* When marker is clicked, location view is shown. When same marker is clicked again
@@ -243,7 +122,7 @@ class App extends Component {
             <div className="map">
                 <GoogleMaps
                   locations={ this.getFilteredLocations(typeFilters) }
-                  overflowLocations={ this.getOverflowLocations(aaltoLocations) }
+                  overflowLocations={ statusLogic.getOverflowLocations(aaltoLocations) }
                   toggleLocationView={ this.toggleLocationView }
                 />
             </div>
